pyproject.toml
[project]
name = "mom-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dateparser>=1.2.2",
    "faster-whisper>=1.2.0",
    "flask>=2.1",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "llama-cpp-python>=0.3.16",
    "pyannote-audio>=3.1.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "tiktoken>=0.12.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "werkzeug>=3.1.3",
    "whisperx>=3.1.1",
]
